Rename file and create test for file
